[REF][l10n_br_nfe] Simplifying the NF-e import wizard
