+++ l10n_br_nfe/security/ir.model.access.csv
@@ -1,8 +1,4 @@
 "id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
-acess_l10n_br_fiscal_import_wizard_user,l10n_br_nfe_import_xml_user,model_l10n_br_nfe_import_xml,l10n_br_nfe.group_user,1,0,0,0
-acess_l10n_br_fiscal_import_wizard_manager,l10n_br_nfe_import_xml_manager,model_l10n_br_nfe_import_xml,l10n_br_nfe.group_manager,1,1,1,1
-acess_l10n_br_fiscal_import_products_wizard_user,l10n_br_nfe_import_xml_products_user,model_l10n_br_nfe_import_xml_products,l10n_br_nfe.group_user,1,0,0,0
-acess_l10n_br_fiscal_import_products_wizard_manager,l10n_br_nfe_import_xml_products_manager,model_l10n_br_nfe_import_xml_products,l10n_br_nfe.group_manager,1,1,1,1
 access_l10n_br_account_product_nfe_export_invoice_user,access_l10n_br_account_product_nfe_export_invoice_user,model_l10n_br_account_product_nfe_export_invoice,l10n_br_nfe.group_user,1,0,0,0
 access_l10n_br_account_product_nfe_export_invoice_manager,access_l10n_br_account_product_nfe_export_invoice_manager,model_l10n_br_account_product_nfe_export_invoice,l10n_br_nfe.group_manager,1,1,1,1
 access_l10n_br_account_product_nfe_export_invoice_result_user,access_l10n_br_account_product_nfe_export_invoice_result_user,model_l10n_br_account_product_nfe_export_invoice_result,l10n_br_nfe.group_user,1,0,0,0
