+++ l10n_br_nfe/views/nfe_menu.xml
@@ -1,23 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
 
-    <menuitem
-        id="nfe_import_document"
-        action="l10n_br_nfe_import_xml_action"
-        name="Import NF-e"
-        parent="l10n_br_fiscal.document_sub_menu"
-        sequence="46"
-        groups="l10n_br_fiscal.group_manager"
-    />
 
-    <!-- Imported Documents Menu-->
-    <menuitem
-        id="imported_document_menu"
-        action="imported_document_action"
-        name="Imported Document"
-        parent="l10n_br_fiscal.document_sub_menu"
-        sequence="20"
-        groups="l10n_br_fiscal.group_user,l10n_br_fiscal.group_manager"
-    />
 
 </odoo>
