+++ l10n_br_nfe/wizards/__init__.py
@@ -1,3 +1,3 @@
 from . import l10n_br_account_nfe_export_invoice
 from . import l10n_br_account_nfe_export
-from . import import_document
+from . import document_import_wizard
